pyproject.toml
[tool.poetry]
name = "oncall"
version = "0.1.0"
description = ""
authors = ["Luke Tchang <ltchang@stanford.edu>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
redis = "^5.2.1"
rq = "^2.1.0"
pydantic = "^2.10.6"
ruff = "^0.9.4"
fastapi = "^0.115.8"
python-multipart = "^0.0.20"
uvicorn = "^0.34.0"
langchain = "^0.3.17"
langchain-community = "^0.3.16"
langchain-openai = "^0.3.3"
sqlalchemy = "^2.0.37"
alembic = "^1.14.1"
psycopg2 = "^2.9.10"
sqlalchemy-utils = "^0.41.2"
langchain-anthropic = "^0.3.5"
python-dotenv = "^1.0.1"
openapi-python-client = "^0.23.1"
langgraph = "^0.2.3"

pytz = "^2025.1"
twilio = "^9.4.4"
phonenumbers = "^8.13.54"
boto3 = "^1.36.13"
pypdf2 = "^3.0.1"
geopy = "^2.4.1"
pytest-mock = "^3.14.0"
pytest = "^8.3.4"
braintrust = "^0.0.184"
autoevals = "^0.0.119"
pytest-asyncio = "^0.25.3"
python-json-logger = "^3.2.1"
discord = "^2.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"